+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- In `ProductSpecifications`, use specificationGroups resolver to get translated values.
 
 ## [3.71.3] - 2019-10-03
 ### Fixed